+++ apps/ui-dev/constants/design-pages.js
@@ -201,7 +201,7 @@ export default [
       },
       {
         link: COMPONENTS_URL + "patterns/table/",
-        title: "Table *",
+        title: "Table",
       },
     ],
     category: "Components",